% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_download.R
\name{bfs_wrapper}
\alias{bfs_wrapper}
\title{A Wrapper for the BFS Packages download function}
\usage{
bfs_wrapper(id, dest_file, do_api_call = FALSE, type = "data")
}
\arguments{
\item{id}{String. The BFS identifier (PX-ID for 'data' or Asset-ID for 'asset').}

\item{dest_file}{String. The local file path where the data should be saved.}

\item{do_api_call}{Bool. If \code{TRUE}, forces a new download even if 
the file exists locally. Defaults to \code{FALSE}.}

\item{type}{String. The type of download: \code{"data"} for API table 
extraction or \code{"asset"} for direct file downloads (e.g., Excel/CSV).}
}
\value{
Bool Returns \code{TRUE} if the file exists or was successfully 
downloaded; \code{FALSE} otherwise.
}
\description{
This wrapper allows to download functions from the BFS Package. Depending on the source
either data or asset we use different functions. This function handles that distinction.
It also only downloads if either teh data isn't at the given path or if there is a forced re-download
}
