% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_formatting.R
\name{clean_stat_string}
\alias{clean_stat_string}
\title{Clean Strings for LaTeX}
\usage{
clean_stat_string(x)
}
\arguments{
\item{x}{A character string of numbers or float.}
}
\value{
The cleaned string
}
\description{
Fixes truncated p-values (e.g., "(0)" -> "(0.00)", "(1)" -> "(1.00)") 
to ensure proper alignment in academic tables.
}
