% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_kof_data.R
\name{download_kof_data_wrapper}
\alias{download_kof_data_wrapper}
\title{Wrapper Function for download_kof_data}
\usage{
download_kof_data_wrapper(
  file,
  kof_data_key = kof_data_key,
  do_api_call = FALSE,
  kof_api_key = NULL
)
}
\arguments{
\item{file}{String. Path to the file, including it's name: "C:/Users/user1/Downloads/filename.csv}

\item{kof_data_key}{String. Reference to what dataset should be downloaded. Found on KOF Webpage.}

\item{do_api_call}{Bool. Set to \code{TRUE} if you want to force an datadownload even if data has already been downloaded}

\item{kof_api_key}{String. KOF API Key if needed.}
}
\value{
None. The function doesn't return anything: It prints messages and downloads directly
}
\description{
This function wraps the download_kof_data to handle the decision wether to call it.
If the file isn't saved at the given path or there is an explicit instruction to download the data
the wrapper skips the download. The wrapper also returns a message on what it does.
}
\examples{
\dontrun{
download_kof_data_wrapper("data/kof_raw.csv", "kof_dataset_123", do_api_call = TRUE, kof_api_key = my_kof_api_key)
}
}
\seealso{
[download_kof_data()]
}
