% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_url_csv}
\alias{download_url_csv}
\title{Download csv from an url}
\usage{
download_url_csv(url, filepath)
}
\arguments{
\item{url}{String. Url where the csv file is}

\item{filepath}{Where the file is stored}
}
\value{
Bool. Messages indicate if download was succesfull and returns \code{TRUE} or \code{False} respectively
}
\description{
This is a very generic function that takes an url and downloads the csv data behind it.
}
\seealso{
[download_url_csv_wrapper()]
}
