% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_download.R
\name{download_url_csv_wrapper}
\alias{download_url_csv_wrapper}
\title{Wrapper Function for the download_url_csv_function}
\usage{
download_url_csv_wrapper(url, filepath, do_api_call)
}
\arguments{
\item{url}{String. Url where the csv file is}

\item{filepath}{Where the file is stored}

\item{do_api_call}{Bool. If \code{TRUE}, forces a download even if the file exists.}
}
\value{
None. Messages indicating success.
}
\description{
Allows to call it only if wither file is missing or there is a forced api call
}
\seealso{
[download_url_csv()]
}
