% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_builder.R
\name{G_matrix_factory}
\alias{G_matrix_factory}
\title{Measurement Link Matrix (G) Factory}
\usage{
G_matrix_factory(Y_data, loadings = NULL, nr = 1)
}
\arguments{
\item{Y_data}{Matrix/DataFrame of observed variables used to determine dimensions.}

\item{loadings}{Numeric vector. Fixed weights for the state's impact on observations. Defaults to 1s.}

\item{nr}{Integer. Number of hidden states, defaults to 1.}
}
\value{
A list containing \code{manifest} (empty for fixed loadings) and \code{builder}.
}
\description{
Factory Function that creates a function
Creates the matrix linking the hidden state (e.g., natural rate) to the observed variables.
It creates a list containing the manifest (parameter specifications) and a
function according to matrix specifications for the optimizer.
}
\seealso{
[ssm_optimizer()]
}
