% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_input_preparation.R
\name{generate_snb_reer_forecasts}
\alias{generate_snb_reer_forecasts}
\title{Generate Rolling Real-Time REER Forecast Matrix for the SNB REER}
\usage{
generate_snb_reer_forecasts(
  data = master_philips,
  date_col = "date",
  start_date = "1990 Q1",
  end_date = NULL,
  forecast_h = 8,
  snb_reer_delay = 4,
  burn_in = "1982 Q1"
)
}
\arguments{
\item{data}{DataFrame containing master historical series.}

\item{date_col}{Character. Column name for target dates in output (default "date").}

\item{start_date}{Character/yearqtr. First forecast origin date (default "1990 Q1").}

\item{end_date}{Character/yearqtr. Last forecast origin date (default NULL -> max date).}

\item{forecast_h}{Integer. Forecast horizon in quarters (default 8).}

\item{snb_reer_delay}{Integer. Lag in quarters before official SNB REER is published (default 3).}

\item{burn_in}{Character/yearqtr. Burn-in start date (default "1982 Q1").}
}
\value{
A data.frame with target dates formatted as YYYYQX in the first column (`date_col`) and vintage origin columns.
}
\description{
Generate Rolling Real-Time REER Forecast Matrix for the SNB REER
}
