% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_input_preparation.R
\name{get_arma_forecast}
\alias{get_arma_forecast}
\title{Generate a single h-step ARMA forecast for a rolling window}
\usage{
get_arma_forecast(
  current_T,
  T_0 = "2000 Q1",
  h = 8,
  data,
  data_col,
  date_col = "quarter"
)
}
\arguments{
\item{current_T}{The forecast origin date (as a `yearqtr` or character string).}

\item{T_0}{The start date for the training sample (default is "2000 Q1").}

\item{h}{Integer. The number of quarters to forecast ahead (default is 8).}

\item{data}{A dataframe or tibble containing the time series.}

\item{data_col}{Character. The name of the column to be forecasted.}

\item{date_col}{Character. The name of the date column (default is "quarter").}
}
\value{
A tibble with `origin_date`, `forecast_date`, `h_step`, and `predicted_value`. 
Returns `NULL` if estimation fails or data is insufficient.
}
\description{
Estimates an ARIMA model from a dataset with a data and a date column to a
specific time origin and returns forecast. The data doesn't need to be transformed
and prepared before running the function, i.e logs and differentiation gets handled in
the function (logs) and by the auto arima (diff). Only works on quarterly data.
}
\details{
The function performs a log-transformation of the input data before estimation:

The ARIMA model is selected automatically and model accounts for non-stationarity
through automatic differencing by the auto.arima function

Forecasts are converted back to levels using the exponential function:
}
