% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_input_preparation.R
\name{get_hp_gap}
\alias{get_hp_gap}
\title{Extract HP Filter Gap at a specific Vantage Point}
\usage{
get_hp_gap(
  data,
  gdp_forecast_data,
  vantage_q,
  h = 8,
  gdp_col = "log_gdp",
  date_col = "quarter",
  return_forecasts = FALSE
)
}
\arguments{
\item{data}{A dataframe containing historical observed data.}

\item{gdp_forecast_data}{A dataframe containing forecasted values for the series extension.}

\item{vantage_q}{The forecast origin date (as a \code{yearqtr} or character string). 
Represents "today" in the pseudo out-of-sample context.}

\item{h}{Integer. The number of forecast quarters to append as an extension. Default is 8.}

\item{gdp_col}{Character. The name of the column containing the series to be filtered (e.g., "log_gdp").}

\item{date_col}{Character. The name of the column containing dates. Default is "quarter".}

\item{return_forecasts}{Logical. If \code{FALSE} (default), returns the historical 
gaps up to \eqn{T}. If \code{TRUE}, returns the filtered gaps for the forecast 
horizon \eqn{T+1} to \eqn{T+h}.}
}
\value{
A tibble containing the date column and the calculated \code{gap}. 
Returns \code{NULL} if the fused series contains fewer than 12 observations.
}
\description{
This function calculates the Hodrick-Prescott (HP) cyclical component (gap) by 
splicing historical observed data untilt valtage point T with forecasted extensions.
This approach mitigates the "end-point bias" of the HP filter in pseudo out-of-sample 
forecasting contexts, and also allows to forecast the Output gap
}
\details{
The function constructs an augmented time series \eqn{Y^*} by combining 
observed data up to the vantage point \eqn{T} and forecasted values up to 
\eqn{T+h}:

The HP filter is then applied to the entire augmented series to extract the 
cycle Component representing the Output gap. if \code{return_forecasts = FALSE}
it returns the oputput gap until the vantage point, where the forecasts serve
as a stabilizer to mitigate endpoint bias, while if its is set to \code{TRUE}
then it returns both historical and the full forecasts of the Output gap up
to horizon \code{T+h}.
}
\seealso{
[build_data_matrix_philips()]
}
