% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_download.R
\name{get_snb_data_wrapper}
\alias{get_snb_data_wrapper}
\title{Wrapper function for download SNB Data}
\usage{
get_snb_data_wrapper(
  file,
  do_api_call = FALSE,
  cube,
  file_name,
  D0 = NULL,
  D1 = NULL,
  D2 = NULL,
  file_type = "csv"
)
}
\arguments{
\item{file}{String. The full path to the local file to check.}

\item{do_api_call}{Bool. If \code{TRUE}, forces a download even if the file exists.}

\item{cube}{String. The SNB cube ID found on their website.}

\item{file_name}{String. Base name for the saved file.}

\item{file_type}{String. Format of the file (default 'csv').}

\item{ids}{(Vector of) Strings. The series ID(s) to fetch.}
}
\value{
Invisibly returns the result of the download attempt.
}
\description{
Wrapper function for the download_snb_data function. It checks if either the data is missing,
or if there is forced download via \code{do_api_call} and downloads if either of these is the case.
}
\seealso{
[download_snb_data()]
}
