% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_forecasting.R
\name{get_ssm_forecast_parameters}
\alias{get_ssm_forecast_parameters}
\title{Generate Rolling SSM Parameters for Forecasting}
\usage{
get_ssm_forecast_parameters(
  data,
  y_cols,
  x_col,
  date_col = "quarter",
  all_builder_functions,
  spec,
  all_defaults,
  forecast_start,
  val_T1 = "2015-01-01",
  forecast_end = NULL
)
}
\arguments{
\item{data}{The Dataframe that contains full data$}

\item{y_cols}{The Name of the Y column(s) in the data dataframe}

\item{x_col}{The column that contains the data for exogenous variales}

\item{date_col}{the column name of the date column}

\item{spec}{the okun model specification}

\item{all_defaults}{the default guesses}

\item{forecast_start}{The date where the forecast stats, limit of the available information set}

\item{val_T1}{the beginning of the forecast information set}
}
\description{
Estimates model parameters over an expanding window up to a specified end date.
Returns a list of parameters to be used in a separate forecasting step.
}
\details{
We have T = 1 as the start of the data, T=T is the end of the data,
T=t as the current date up until we have the data availabe for the pseudo forecast
and t+h would then be the forecast horizon.
What we do here is just get the parameter estimation at time t
}
