% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_input_preparation.R
\name{get_ssm_forecasted_inf_gap}
\alias{get_ssm_forecasted_inf_gap}
\title{Build the INf gap using the structural rate of inflation}
\usage{
get_ssm_forecasted_inf_gap(inf_forecasts_df)
}
\description{
NOT YET IMPLEMENTED
}
