% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_builder.R
\name{H_matrix_factory}
\alias{H_matrix_factory}
\title{Transition Matrix (H) Factory
Factory Function that creates a function
This Function Creates the H matrix of for the Kalmann filter, which gouverns
the persistance of the state. It can be either an AR(1) Process (default) or
a Randowm walk when there is no persistance parameter to estimate and the
matrix defaults to 1.}
\usage{
H_matrix_factory(name = "phi", random_walk = FALSE)
}
\arguments{
\item{name}{give the AR(1) parameter a name, defaults to phi}

\item{random_walk}{If set to true, the phi parameter is set to 1 for a random walk}
}
\value{
returns a list with 2 objects, one is the matrix function used for the optimizer
and the parameter rule that translates between optimizer and the model, as the AR(1) param
has to be between ]0;1[
}
\description{
Transition Matrix (H) Factory
Factory Function that creates a function
This Function Creates the H matrix of for the Kalmann filter, which gouverns
the persistance of the state. It can be either an AR(1) Process (default) or
a Randowm walk when there is no persistance parameter to estimate and the
matrix defaults to 1.
}
\seealso{
[loglik_ssm()]
}
