% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model_matrices.R
\name{initialize_my_philips_ssm}
\alias{initialize_my_philips_ssm}
\title{Initialize the Phillips Curve State-Space Model Blueprint}
\usage{
initialize_my_philips_ssm(Y_data, X_data, parameter_guesses)
}
\arguments{
\item{Y_data}{A matrix or dataframe containing the observation variables (typically 
actual inflation and inflation expectations).}

\item{X_data}{A matrix or dataframe containing the exogenous drivers (typically 
lagged inflation, output gap(s), and LOP gap(s)).}

\item{parameter_guesses}{A named list containing initial starting values for 
\code{beta1}, \code{beta2}, \code{beta3}, \code{sigma_cpi}, and \code{sigma_spf}.
Optional parameters when left out will used predetermined fixed values.}
}
\value{
A structured list (\code{ssm}) containing:
\itemize{
  \item \code{data}: A list with matrices \code{Y} and \code{X}.
  \item \code{manifest}: A list defining the initial values and transformation rules for each parameter.
  \item \code{builders}: A list of function pointers (\code{mu_t}, \code{H}, \code{G}, \code{M}, \code{N}) 
  used to construct the SSM matrices during filtering.
}
}
\description{
This function initializes a structured State-Space Model (SSM) object
specifically for the Phillips curve estimation. It defines the parameter 
manifest, containing what parameters are to be estimated and what the initial guesses are,
the functions that build the matrices for the optimizer and the used data.
}
\details{
The manifest also sets the specific constraints on parameter estimation:
\itemize{
  \item \bold{Rule 0 (Linear):} No transformation. Applied to \code{beta1} (lagged inflation), 
  \code{beta2} (output gap), and \code{beta3} (exchange rate gap).
  \item \bold{Rule 1 (Exponential):} Constraints parameters to be strictly positive (\eqn{>0}). 
  Applied to measurement noise (\code{sigma_cpi}, \code{sigma_spf}) and state variance (\code{xi_n}).
  \item \bold{Rule 2 (Logit):} Constraints parameters between 0 and 1. Applied to 
  state persistence (\code{phi}).
}

If \code{phi} is not provided, the model defaults to a Random Walk (\eqn{\phi = 1}) for the State Transition.
If \code{xi_n} is not provided, the model utilizes a fixed variance defined within 
the \code{build_N} function for the state variance
}
\seealso{
\code{\link{build_mu_t}} for exogenous regressor matrix construction.

\code{\link{build_H}} for state factor loading matrices.

\code{\link{build_G}} for state-to-measurement mapping.

\code{\link{build_M}} for measurement error covariance structures.

\code{\link{build_N}} for state innovation covariance matrices.
}
