% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_builder.R
\name{M_matrix_factory}
\alias{M_matrix_factory}
\title{Measurement Noise Matrix (M) Factory}
\usage{
M_matrix_factory(Y_data, prefix = "sigma_")
}
\arguments{
\item{Y_data}{Matrix/DataFrame of observations to determine the number of variances.}

\item{prefix}{Character. Prefix for the variance parameters, defaults to "sigma_".}
}
\value{
A list containing \code{manifest} (Rule 1: Exponential to be non negative) and \code{builder}.
}
\description{
Factory Function that creates a function
Creates a diagonal covariance matrix for observation errors. The parameter specs
are restricted to be positive in the manifest (parameter specs)
}
