% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mround}
\alias{mround}
\title{Round Numbers to the Nearest Multiple}
\usage{
mround(x, step = 0.25)
}
\arguments{
\item{x}{Numeric vector to be rounded.}

\item{step}{Numeric scalar representing the rounding increment/multiple. Defaults to \code{0.25}.}
}
\value{
A numeric vector rounded to the nearest multiple of \code{step}.
}
\description{
Rounds numeric values to the nearest specified step or increment (e.g., nearest 0.25).
}
