% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_builder.R
\name{mu_t_matrix_factory}
\alias{mu_t_matrix_factory}
\title{Exogenous Component Matrix (mu_t) Factory}
\usage{
mu_t_matrix_factory(
  X_data,
  Y_data,
  param_prefix = "beta",
  intercept = FALSE,
  impact_cols = c(1)
)
}
\arguments{
\item{X_data}{Matrix of exogenous regressors}

\item{Y_data}{Matrix of observations to determine time and variable dimensions.}

\item{param_prefix}{Character. Prefix for the coefficients, defaults to "beta".}

\item{intercept}{Bool. Whether to include an estimated intercept.}

\item{impact_cols}{Integer vector. Indices of observation variables influenced by the signal.}
}
\value{
A list containing \code{manifest} (Rule 0: Linear) and \code{builder}.
}
\description{
Factory Function that creates a function.
Creates a time-varying matrix of exogenous signals. Can do this using exogenous inputs.
}
