% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_forecasting.R
\name{predict_ssm_path_rw}
\alias{predict_ssm_path_rw}
\title{Function to predict the random walk ssm with exogenous betas}
\usage{
predict_ssm_path_rw(rho_start, betas, exog_features)
}
\arguments{
\item{betas}{the estimated beta coefficients as dataframe or matrix}

\item{exog_features}{exogenous features corresponding to betas as matrix or dataframe

This function takes the last state variable and adds the exogenous component
built from using a matrix multiplication of exogenous features and betas and
adding the state.

gdp features is a data frame where each row is a time in h and each column is a lag}

\item{who_start}{the last state observation from the estimation}
}
\description{
Function to predict the random walk ssm with exogenous betas
}
