% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_download.R
\name{snb_api_data_to_ts}
\alias{snb_api_data_to_ts}
\title{Format SNB Time Series}
\usage{
snb_api_data_to_ts(
  dataset,
  variable,
  variable_name,
  names_column = "D0",
  values_column = "Value"
)
}
\arguments{
\item{dataset}{Dataframe. The raw data downloaded from the SNB API.}

\item{variable}{Sting. The ID of the specific variable to extract from the D0 column.}

\item{variable_name}{String. The new name to assign to the value column.}

\item{names_column}{String. The column containing variable IDs. Default is "D0".}

\item{values_column}{String. The column containing the numeric values. Default is "Value".}
}
\value{
A processed dataframe with a \code{date} column and the renamed value column.
}
\description{
Function that prepares SNB Data to get transformed with the format_time_series_df function, then applies it.
}
