% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_input_preparation.R
\name{splice_reer_series}
\alias{splice_reer_series}
\title{Splice Official SNB REER with CREA Proxy and Component Forecasts}
\usage{
splice_reer_series(
  vantage_quarter = "2023 Q2",
  snb_reer_delay = 4,
  data = master_philips,
  burn_in = "2010 Q1"
)
}
\arguments{
\item{vantage_quarter}{The date of the forecast origin.}

\item{snb_reer_delay}{Number of quarters SNB data is assumed to be lagging.}

\item{data}{Master quarterly dataframe.}

\item{burn_in}{Start date for the series history.}
}
\value{
A tibble with [quarter, reer_simulated] extended h-steps ahead.
}
\description{
Splice Official SNB REER with CREA Proxy and Component Forecasts
}
\details{
To run the pseudo out of sample forecasts correctly we need to take into account
that the SNB REER releases with a delay of 1 year. Therefore we need to remove
that timeframe from the data and replace it with a proxy. Due to teh release delay
of data in reality the delay is usually only 3 quarters, as GDP for example releases
in the middle of the next quarter, at whch point the SNB data is already here.
To be conservative however we still use 4 quarters
}
