% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_implementation.R
\name{calculate_kalman_gain}
\alias{calculate_kalman_gain}
\title{Calculate Kalman Gain}
\usage{
calculate_kalman_gain(sigma, G, R)
}
\arguments{
\item{sigma}{State estimation covariance matrix (Sigma_{t|t-1}).}

\item{G}{Observation/Mapping matrix.}

\item{R}{Observation noise covariance matrix.}
}
\value{
A matrix representing the Kalman gain.
}
\description{
Computes the optimal Kalman gain matrix (K) for the state update step.
}
