% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_download.R
\name{download_kof_data}
\alias{download_kof_data}
\title{Download Data from the KOF Institute at ETHZ}
\usage{
download_kof_data(
  file_path = kof_master,
  dataset = kof_data_key,
  api_key = NULL
)
}
\arguments{
\item{file_path}{String. Where to save the dataset. Needs to include the actual file name such as "C:/Users/user1/Downloads/filename.csv}

\item{dataset}{String. Reference to what dataset should be downloaded. Found on KOF Webpage.}

\item{api_key}{String. If a KOF API Key is needed, add here. \code{NULL} by default}
}
\value{
Bool. Returns \code{TRUE} if download was successful and \code{FALSE} if it wasn't.
}
\description{
This function calls on the KOF package to download a dataset to a certain
save location. It transforms the downloaded list of ts objects into a data
frame for easier saving and use.
}
\seealso{
[download_kof_data_wrapper()]

[https://kof.ethz.ch/]
}
