% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_snb_data.R
\name{download_snb_data}
\alias{download_snb_data}
\title{Function to Download Data from SNB API}
\usage{
download_snb_data(
  cube = "zimoma",
  folder = "data",
  file_name = "name",
  from_date = "2010-01",
  to_date = "2026-01",
  D0 = c("SARON", "3M0"),
  D1 = NULL,
  D2 = NULL,
  file_type = "csv"
)
}
\arguments{
\item{cube}{String. The specific data set identifier from the SNB website (e.g., 'zimoma').}

\item{folder}{String. The storage directory. Must exist prior to calling the function.}

\item{file_name}{String. The base name for the saved files.}

\item{from_date}{String. Start date in 'YYYY-MM' format.}

\item{to_date}{String. End date in 'YYYY-MM' format.}

\item{D0}{(Vector of) Strings. One or more series identifiers (e.g., 'SARON'), D0 in the cube.}

\item{D1}{(Vector of) Strings. One or more series identifiers. Usually subcategories of D0 series. D1 in the cube}

\item{D2}{(Vector of) Strings. One or more series identifiers. Usually subcategories of D0, D1, series. D2 in the cube.}

\item{file_type}{String. The file format for the data download. Default is 'csv'.}
}
\value{
Bool. An indicator: \code{TRUE} if download was successful \code{FALSE} otherwise.
}
\description{
This function transforms the R script on the SNB's website into a reusable function to download Data.
An example dataset is available here: 
It does download the Data directly, it does not return the Data
}
\seealso{
[get_snb_data_wrapper()]

[https://data.snb.ch/en/topics/ziredev/cube/zimoma]

[https://data.snb.ch/en]
}
