% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_forecasting.R
\name{forecast_okun_ssm}
\alias{forecast_okun_ssm}
\title{This function runs the actual forecast for the okun model}
\usage{
forecast_okun_ssm(
  params_df = okun_params_df,
  date_col = "quarter",
  exog_var_col = "log_gdp",
  forecast_h = 8,
  Y_data_df = df_okun_final,
  target_variable = "unemp_rate",
  X_data = X_okun,
  gdp_gap_forecasts = gdp_gap_forecasts
)
}
\arguments{
\item{params_df}{a dataframe containing all necessary parameters for forecsting
and the final state estimate}

\item{date_col}{the name of the date column in the data_df}

\item{exog_var_col}{name of the exogenous fata (gdp) column}

\item{forecast_h}{forecast horizon, how many quarters ahead we should forecast}

\item{Y_data_df}{the dataframe containing dependent Y observed data}

\item{target_variable}{name of the dependent variable that we want to forecast}

\item{X_data}{Independent regressors added to the data}

\item{gdp_gap_forecasts}{External forecasts that are used as a forecast in the external regressors}
}
\value{
A dataframe whith the forecats of the selected target variable. The fdiagonal contains the
true observed variables, the values below are the forecats at the column date for the rowdate

The goal of the function is a forecast of the unemployment rate using an Okun model,
where the gdp gap acts as an external regressor. This function uses as input a parameter df
which contains estimated up to a certain horizon for a a series of dates. then the function loops over all these dates
and runs a pseudo out of sample forecast.

The difficulty in this function is the handling of the external regressors. At each date of the pseudo forecast,
(vantage point) we have to rerun the HP filter to get the cycles we'd have had at that time. To reliably
forecast the GDP gap we also use an forecast for the next h periods of gdp that we fuse together with the observed gdp
and then run the filter. We also use a burn in for the HP filter. This means that the estimation window starts in 2015 for this model
we start calculating the HP filter in 2010.
}
\description{
This function runs the actual forecast for the okun model
}
\seealso{
[get_ssm_forecast_parameters()]

[predict_ssm_path_rw()]
}
