% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_implementation.R
\name{kalman_filter}
\alias{kalman_filter}
\title{Constrained Kalman Filter for Linear State-Space Models}
\usage{
kalman_filter(
  Y_t,
  nu_t,
  H,
  N,
  mu_t,
  G,
  M,
  Sigma_0,
  rho_0,
  indic_pos = 0,
  Rfunction = calc_covariance,
  Qfunction = calc_covariance,
  reconciliationf = function(x, opt) {
     x
 }
)
}
\arguments{
\item{Y_t}{Matrix. Observed data matrix of dimension \eqn{T \times ny}.}

\item{nu_t}{Matrix. Transition equation intercept/exogenous component (\eqn{T \times nr}). 
Depends on past information only.}

\item{H}{Matrix. Transition matrix describing the evolution of the state (e.g., AR(1) coefficients).}

\item{N}{Matrix. State noise scaling matrix (process noise).}

\item{mu_t}{Matrix. Measurement intercept matrix (\eqn{T \times ny}). Would be 0 if data is demeaned.}

\item{G}{Matrix. Observation matrix containing the relations between observed and unobserved variables.}

\item{M}{Matrix. Measurement noise scaling matrix.}

\item{Sigma_0}{Matrix. Initial covariance matrix guess for the state vector.}

\item{rho_0}{Vector. Starting guess for the initial state vector.}

\item{indic_pos}{Vector. Binary indicators (0/1) specifying if specific latent factors 
must be constrained to be \eqn{\ge 0}.}

\item{Rfunction}{Function. Defines the measurement noise covariance. Defaults to \code{calc_covariance}.}

\item{Qfunction}{Function. Defines the transition noise covariance. Defaults to \code{calc_covariance}.}

\item{reconciliationf}{Function. Define a modification for \code{rho_tt} (e.g., for consistency 
in Augmented State Vectors in Quadratic Kalman Filters).}
}
\value{
A list containing:
\item{r}{Filtered states \eqn{\rho_{t|t}} (size \eqn{T \times nr}).}
\item{Sigma_tt}{Filtered state covariance \eqn{\Sigma_{t|t}} stored as flattened vectors.}
\item{loglik}{Scalar. Total log-likelihood of the model.}
\item{y_tp1_t}{Forecasted observables \eqn{y_{t+1|t}}.}
\item{S_tp1_t}{Predicted state covariance \eqn{\Sigma_{t+1|t}}.}
\item{r_tp1_t}{Predicted state vector \eqn{\rho_{t+1|t}}.}
\item{loglik.vector}{Vector of date-specific log-likelihoods.}
\item{Omega_tp1_t}{Innovation covariance \eqn{\Omega_{t+1|t}}.}
\item{M}{The measurement scaling matrix used.}
\item{fitted_obs}{Estimated observables given filtered states (\eqn{\hat{y}_t|t}).}
}
\description{
Implements a Kalman filter for the estimation of latent states in a linear system.
This implementation supports exogenous components, missing values (NA), 
non-negativity constraints on latent factors, and potential quadratic reconciliation.
}
\note{
If \code{reconciliationf} is provided, \code{rho_tt} is modified to ensure consistency 
defined by the function (used in QKF for augmented state vectors).
}
\section{System Equations}{

Measurement equations:
\deqn{y_t = \mu_t + G \rho_t + M \epsilon_t}
Transition equations:
\deqn{\rho_t = \nu_t + H \rho_{t-1} + N \xi_t}
}

