% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_builder.R
\name{loglik_ssm}
\alias{loglik_ssm}
\title{State-Space Log-Likelihood Function}
\usage{
loglik_ssm(
  theta,
  Y,
  model_spec,
  mu_t_builder,
  G_builder,
  H_builder,
  M_builder,
  return_full_res = FALSE
)
}
\arguments{
\item{theta}{Numeric vector of parameters furnished by optimizer}

\item{Y}{Observed data matrix.}

\item{model_spec}{Parameter transformation dictionary containing restrictions}

\item{mu_t_builder}{Builder function for the exogenous component matrix.}

\item{G_builder}{Builder function  for the measurement link matrix.}

\item{H_builder}{Builder function  for the transition matrix.}

\item{M_builder}{Builder function  for the measurement noise matrix.}
}
\value{
Numeric. The negative log-likelihood.
}
\description{
Calculates the negative log-likelihood of a State-Space model given current parameters.
Integrates parameter mapping, matrix building, and Kalman filtering.
}
