% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_builder.R
\name{model2param_gen}
\alias{model2param_gen}
\title{Map Model Parameters to Optimizer Parameters}
\usage{
model2param_gen(input_list, spec)
}
\arguments{
\item{spec}{List. Transformation rules.}

\item{model_list}{Named list of model parameters.}
}
\value{
A numeric vector of unconstrained parameters.
}
\description{
Inverse of \code{param2model_gen}. Maps model values back to unconstrained
space for the optimizer.
}
\seealso{
[param2model_gen()]
}
