% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_builder.R
\name{param2model_gen}
\alias{param2model_gen}
\title{Map Optimizer Parameters to Model Parameters}
\usage{
param2model_gen(theta, spec)
}
\arguments{
\item{theta}{Numeric vector. Unconstrained parameters from the optimizer.}

\item{spec}{List. Contains parameter names and their transformation rules.}
}
\value{
A named list of transformed model parameters.
}
\description{
Transforms unconstrained values from the optimizer (that gives random values)
into model-appropriate values according to parameter restictions using specified
rules (e.g., exponential for variance or non negative parameters, or logistic
for persistence that are between 0 and 1). These are then transformed and plugged
into the matrix functions
}
\seealso{
[model2param_gen()]
}
