% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_builder.R
\name{rolling_est_okun_ssm}
\alias{rolling_est_okun_ssm}
\title{Rolling Okun SSM Estimation}
\usage{
rolling_est_okun_ssm(
  data,
  forecast_start,
  forecast_end = NULL,
  date_col = "quarter",
  val_T1 = "2015-01-01"
)
}
\arguments{
\item{data}{Raw merged dataframe containing log_gdp, unemp_rate, and spf_5y_unemp}

\item{forecast_start}{The first date to generate a forecast/parameter set for}

\item{forecast_end}{End date of rolling window (optional)}

\item{val_T1}{The start date for the Kalman Filter (e.g., 2015-01-01)}
}
\description{
Generate Rolling SSM Parameters for Forecasting
}
\details{
Estimates model parameters over an expanding window up to a specified end date.
Returns a list of parameters to be used in a separate forecasting step.

We have T = 1 as the start of the data, T=T is the end of the data,
T=t as the current date up until we have the data availabe for the pseudo forecast
and t+h would then be the forecast horizon.
What we do here is just get the parameter estimation at time t.

At each step we have to reestimate the HP filter to get the gaps. This is done via a burn in
where we start the estimation at the first observation in the dataframe. the dataframe is cut to the burn in start in the preparation
this assures there is no starting point bias.
}
