% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{try_install}
\alias{try_install}
\title{Helper function to try installing packages from binary or source if that fails}
\usage{
try_install(pkg)
}
\arguments{
\item{pkg}{Character the package to be installed}
}
\value{
None if success, stops process if error
}
\description{
Helper function to try installing packages from binary or source if that fails
}
\seealso{
[install_missing_packages()]
}
